moved functions to the date_utils file
